Rename interactors, use modules instead of namespaces in class names
