+++ app/controllers/problems_controller.rb
@@ -8,7 +8,7 @@ class ProblemsController < ApplicationController
 
   def show
     @comments = @problem.comments.order(:created_at)
-    View::Create.call(user: current_user, problem: @problem)
+    Views::Create.call(user: current_user, problem: @problem)
   end
 
   def new
