+++ app/interactors/views/create.rb
@@ -1,0 +1,9 @@
+module Views
+  class Create
+    include Interactor
+
+    def call
+      View.where(problem: context.problem, user: context.user).first_or_create
+    end
+  end
+end
